Window-Check throws exception if window is not there. (#284)
* Update index.js

A check with a plain `window` does break if window is not there. If window is undefined e.g. in React Native it will throw an Uncaught ReferenceError: window is not defined(…). This check must be written with a typeof. 

But a problem remains. The fallback to localhost will be wrong in the case of react native.

* Update auth.js

Same here :-)

`!window` does not work as intented. Rewrote to `typeof window !== 'object'`

`window.localStorage === undefined` works because window is at this point a object and returns undefined instead of an exception.